+++ CHANGELOG.md
@@ -11,9 +11,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 - Add `--overwrite` arguments to `facile-rs bag create` and `facile-rs bagpack create` to overwrite existing `--bag-path`.
 - Rename `--datacite-path` to `--datacite-location` in `facile-rs bagpack create` and allow for remote locations.
 - The "publisher" field in Zenodo metadata is now populated with the "publisher" name provided in the Codemeta file.
+- Include MathJax to HTML outputs in `run_docstring_pipeline`, and add option `--mathjax-location` to customize MathJax location or disable it.
 
 ### Changed
 - The option `--overwrite` is not read from environment anymore: it shall be passed from the command line.
+- When HTML output is selected in run_docstring_pipeline, add Mathjax script for rendering math.
 
 ### Fixed
 - Fix remove-doubles when name is given as givenName and familyName.